refactor(lib): extract lib/dhcp.php (ADR-004, A25) — closes #914
