+++ Simple-PHP-IPAM/lib.php
@@ -6,6 +6,7 @@ require_once __DIR__ . '/lib/ip.php';
 require_once __DIR__ . '/lib/config.php';
 require_once __DIR__ . '/lib/db.php';
 require_once __DIR__ . '/lib/demo_seed.php'; // demo-data fixture seeder (ADR-004, A14, #909)
+require_once __DIR__ . '/lib/dhcp.php';      // DHCP config renderers (ADR-004, A25, #914)
 require_once __DIR__ . '/lib/audit.php';
 require_once __DIR__ . '/lib/presentation.php';
 require_once __DIR__ . '/lib/settings.php';
@@ -6333,262 +6334,6 @@ function ipam_send_email_verification(PDO $db, int $userId, string $newEmail): a
     }
 
     return ['success' => true, 'error' => ''];
-}
-
-// ---------------------------------------------------------------------------
-// DHCP config renderers — v3.4.0 #403
-// ---------------------------------------------------------------------------
-
-/**
- * Render an ISC dhcpd.conf snippet for the given (or all) IPv4 subnets.
- *
- * @param int[] $subnetIds Leave empty to include all IPv4 subnets.
- */
-function ipam_render_dhcpd_conf(PDO $db, array $subnetIds): string
-{
-    require_once __DIR__ . '/version.php';
-    $subnets = ipam_dhcp_load_subnets($db, $subnetIds);
-    $lines   = [
-        '# Generated by Simple PHP IPAM v' . IPAM_VERSION
-            . ' on ' . gmdate('Y-m-d H:i:s') . ' UTC',
-        '# Subnets: ' . count($subnets),
-        '',
-    ];
-
-    foreach ($subnets as $s) {
-        $netmask = ipam_prefix_to_netmask(to_int($s['prefix']));
-        $lines[] = 'subnet ' . to_str($s['network']) . ' netmask ' . $netmask . ' {';
-
-        if (!empty($s['dhcp_routers'])) {
-            $safeIps = implode(', ', array_filter(
-                array_map('trim', explode(',', to_str($s['dhcp_routers']))),
-                fn($t) => $t !== '' && filter_var($t, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4) !== false
-            ));
-            if ($safeIps !== '') $lines[] = '  option routers ' . $safeIps . ';';
-        }
-        if (!empty($s['dhcp_dns_servers'])) {
-            $safeIps = implode(', ', array_filter(
-                array_map('trim', explode(',', to_str($s['dhcp_dns_servers']))),
-                fn($t) => $t !== '' && filter_var($t, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4) !== false
-            ));
-            if ($safeIps !== '') $lines[] = '  option domain-name-servers ' . $safeIps . ';';
-        }
-        if (!empty($s['dhcp_domain_name'])) {
-            $domain = trim(to_str($s['dhcp_domain_name']));
-            if (preg_match('/^[a-zA-Z0-9]([a-zA-Z0-9\-.]{0,253}[a-zA-Z0-9])?$/', $domain)) {
-                $quoted  = str_replace(['\\', '"'], ['\\\\', '\\"'], $domain);
-                $lines[] = '  option domain-name "' . $quoted . '";';
-            }
-        }
-        if ($s['dhcp_lease_default'] !== null && $s['dhcp_lease_default'] !== '') {
-            $lines[] = '  default-lease-time ' . to_int($s['dhcp_lease_default']) . ';';
-        }
-        if ($s['dhcp_lease_max'] !== null && $s['dhcp_lease_max'] !== '') {
-            $lines[] = '  max-lease-time ' . to_int($s['dhcp_lease_max']) . ';';
-        }
-        if (!empty($s['dhcp_next_server'])) {
-            $nextSrv = trim(to_str($s['dhcp_next_server']));
-            if (filter_var($nextSrv, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4) !== false) {
-                $lines[] = '  next-server ' . $nextSrv . ';';
-            }
-        }
-        if (!empty($s['dhcp_boot_filename'])) {
-            $bootFile = preg_replace('/[;\n\r{}]/', '', trim(to_str($s['dhcp_boot_filename']))) ?? '';
-            if ($bootFile !== '') {
-                $quoted  = str_replace(['\\', '"'], ['\\\\', '\\"'], $bootFile);
-                $lines[] = '  filename "' . $quoted . '";';
-            }
-        }
-
-        foreach (ipam_dhcp_load_reservations($db, to_int($s['id'])) as $r) {
-            $mac  = ipam_normalize_mac_for_dhcp(to_str($r['mac']));
-            if ($mac === null) continue;
-            $rawHost = to_str($r['hostname']);
-            $label   = ipam_dhcp_normalize_hostname($rawHost) ?? 'host';
-            $suffix  = str_replace('.', '-', to_str($r['ip']));
-            $name    = $label . '-' . $suffix;
-            $lines[] = '  host ' . $name . ' {';
-            $lines[] = '    hardware ethernet ' . $mac . ';';
-            $lines[] = '    fixed-address ' . to_str($r['ip']) . ';';
-            $lines[] = '  }';
-        }
-
-        $lines[] = '}';
-        $lines[] = '';
-    }
-
-    return implode("\n", $lines);
-}
-
-/**
- * Render a Kea 2.x JSON config fragment for the given (or all) IPv4 subnets.
- *
- * @param int[] $subnetIds Leave empty to include all IPv4 subnets.
- */
-function ipam_render_kea_json(PDO $db, array $subnetIds): string
-{
-    $subnets = ipam_dhcp_load_subnets($db, $subnetIds);
-    $subnet4 = [];
-
-    foreach ($subnets as $s) {
-        /** @var array<string,mixed> $entry */
-        $entry = ['subnet' => to_str($s['cidr'])];
-
-        $optionData = [];
-        if (!empty($s['dhcp_routers'])) {
-            $safeRouters = implode(',', array_filter(
-                array_map('trim', explode(',', to_str($s['dhcp_routers']))),
-                fn($t) => $t !== '' && filter_var($t, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4) !== false
-            ));
-            if ($safeRouters !== '') $optionData[] = ['name' => 'routers', 'data' => $safeRouters];
-        }
-        if (!empty($s['dhcp_dns_servers'])) {
-            $safeDns = implode(',', array_filter(
-                array_map('trim', explode(',', to_str($s['dhcp_dns_servers']))),
-                fn($t) => $t !== '' && filter_var($t, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4) !== false
-            ));
-            if ($safeDns !== '') $optionData[] = ['name' => 'domain-name-servers', 'data' => $safeDns];
-        }
-        if (!empty($s['dhcp_domain_name'])) {
-            $optionData[] = ['name' => 'domain-name', 'data' => trim(to_str($s['dhcp_domain_name']))];
-        }
-        if (!empty($s['dhcp_next_server'])) {
-            $optionData[] = ['name' => 'tftp-server-name', 'data' => trim(to_str($s['dhcp_next_server']))];
-        }
-        if (!empty($s['dhcp_boot_filename'])) {
-            $optionData[] = ['name' => 'boot-file-name', 'data' => trim(to_str($s['dhcp_boot_filename']))];
-        }
-        if (!empty($optionData)) {
-            $entry['option-data'] = $optionData;
-        }
-
-        if ($s['dhcp_lease_default'] !== null && $s['dhcp_lease_default'] !== '') {
-            $entry['valid-lifetime'] = to_int($s['dhcp_lease_default']);
-        }
-        if ($s['dhcp_lease_max'] !== null && $s['dhcp_lease_max'] !== '') {
-            $entry['max-valid-lifetime'] = to_int($s['dhcp_lease_max']);
-        }
-
-        $reservations = ipam_dhcp_load_reservations($db, to_int($s['id']));
-        if (!empty($reservations)) {
-            $resArr = [];
-            foreach ($reservations as $r) {
-                $mac = ipam_normalize_mac_for_dhcp(to_str($r['mac']));
-                if ($mac === null) continue;
-                /** @var array<string,mixed> $resEntry */
-                $resEntry = [
-                    'hw-address' => $mac,
-                    'ip-address' => to_str($r['ip']),
-                ];
-                $label = ipam_dhcp_normalize_hostname(to_str($r['hostname']));
-                if ($label !== null) {
-                    $resEntry['hostname'] = $label;
-                }
-                $resArr[] = $resEntry;
-            }
-            $entry['reservations'] = $resArr;
-        }
-
-        $subnet4[] = $entry;
-    }
-
-    $doc = ['Dhcp4' => ['subnet4' => $subnet4]];
-    $json = json_encode($doc, JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_INVALID_UTF8_SUBSTITUTE);
-    if ($json === false) throw new RuntimeException('json_encode failed: ' . json_last_error_msg());
-    return $json;
-}
-
-/**
- * Load IPv4 subnets with DHCP option columns, ordered by network_bin.
- *
- * @param int[] $subnetIds If non-empty, restrict to these IDs.
- * @return list<array<string,mixed>>
- */
-function ipam_dhcp_load_subnets(PDO $db, array $subnetIds): array
-{
-    $sql = "SELECT id, cidr, network, prefix,
-                   dhcp_routers, dhcp_dns_servers, dhcp_domain_name,
-                   dhcp_lease_default, dhcp_lease_max,
-                   dhcp_next_server, dhcp_boot_filename
-            FROM subnets
-            WHERE ip_version = 4";
-
-    if (empty($subnetIds)) {
-        $st = $db->query($sql . " ORDER BY network_bin");
-        if ($st === false) return [];
-        /** @var list<array<string,mixed>> */
-        return $st->fetchAll();
-    }
-
-    $placeholders = implode(',', array_fill(0, count($subnetIds), '?'));
-    $st = $db->prepare($sql . " AND id IN ({$placeholders}) ORDER BY network_bin");
-    $st->execute(array_values(array_map('intval', $subnetIds)));
-    /** @var list<array<string,mixed>> */
-    return $st->fetchAll();
-}
-
-/**
- * Load reserved addresses with a non-empty MAC for a subnet.
- *
- * @return list<array<string,mixed>>
- */
-function ipam_dhcp_load_reservations(PDO $db, int $subnetId): array
-{
-    $st = $db->prepare(
-        "SELECT ip, mac, hostname FROM addresses
-         WHERE subnet_id = :sid AND status = 'reserved' AND mac != ''
-         ORDER BY ip_bin"
-    );
-    $st->execute([':sid' => $subnetId]);
-    /** @var list<array<string,mixed>> */
-    return $st->fetchAll();
-}
-
-/** Convert a CIDR prefix length to a dotted-decimal netmask string. */
-function ipam_prefix_to_netmask(int $prefix): string
-{
-    if ($prefix === 0) return '0.0.0.0';
-    $mask = (int)((0xFFFFFFFF << (32 - $prefix)) & 0xFFFFFFFF);
-    return (string)(inet_ntop(pack('N', $mask)) ?: '0.0.0.0');
-}
-
-/** Normalise any MAC address format to colon-separated lowercase octets. */
-function ipam_normalize_mac_for_dhcp(string $mac): ?string
-{
-    $hex = preg_replace('/[^0-9a-fA-F]/', '', $mac);
-    if (!is_string($hex) || strlen($hex) !== 12) return null;
-    return implode(':', str_split(strtolower($hex), 2));
-}
-
-/**
- * Normalise a DHCP reservation hostname for both Kea and ISC dhcpd output.
- * Returns null if no usable label can be derived; caller falls back to
- * 'host' synthesis.
- *
- * Single-label, RFC 1123 letter-digit-hyphen, leading alpha (after trimming
- * leading digits/hyphens), max 63 chars per label. Dotted inputs take the
- * leftmost label — multi-label FQDNs are out of scope; neither current
- * renderer supports them well.
- *
- * (#1163, PASS-C F-S6-01) — both renderers must agree on what's emitted.
- */
-function ipam_dhcp_normalize_hostname(string $raw): ?string
-{
-    $raw = trim($raw);
-    if ($raw === '') return null;
-    $label = explode('.', $raw, 2)[0];
-    $label = preg_replace('/[^a-zA-Z0-9-]/', '-', $label) ?? '';
-    $label = ltrim($label, '0123456789-');
-    $label = rtrim($label, '-');
-    if ($label === '') return null;
-    if (strlen($label) > 63) {
-        $label = substr($label, 0, 63);
-        // Truncation can land on a hyphen; re-trim so the emitted label
-        // still conforms to RFC 1123 (no trailing '-').
-        $label = rtrim($label, '-');
-        if ($label === '') return null;
-    }
-    return $label;
 }
 
 // ── Custom field definitions (v3.5.0, #313/#596) ──────────────────────────
